Fixing getWitness address normalization bugs in Local state provider
