+++ packages/wallet/core/src/state/utils.ts
@@ -48,3 +48,12 @@ async function retrieveWallets<S extends Signer | SapientSigner>(
     return stateReader.getWallets(await signer.address) as unknown as any
   }
 }
+
+export function normalizeAddressKeys<T extends Record<string, unknown>>(obj: T): Record<Address.Address, T[keyof T]> {
+  return Object.fromEntries(
+    Object.entries(obj).map(([wallet, signature]) => {
+      const checksumAddress = Address.checksum(wallet)
+      return [checksumAddress, signature]
+    }),
+  ) as Record<Address.Address, T[keyof T]>
+}
